+++ src/components/Create/CreateUI/InterpolateUI/InterpolateUI.tsx
@@ -8,7 +8,7 @@ import { Button, Form, Input, InputNumber, Space } from 'antd'
 import axios from 'axios'
 
 // COMPONENTS
-import VideoResult from '@/components/VideoResult/VideoResult'
+import VideoResult from '@/components/Media/VideoResult/VideoResult'
 
 interface InterpolateFormInputs {
   prompt1: string
